requirements.txt
streamlit
pandas
numpy
scikit-learn
joblib
google-generativeai
